FIXUP: quic_max_int: looks like it should be uint64_t
include/haproxy/xprt_quic.h: In function 'quic_max_int':
include/haproxy/xprt_quic.h:317:3: error: large integer implicitly truncated to unsigned type [-Werror=overflow]

+++ include/haproxy/xprt_quic.h
@@ -304,7 +304,7 @@ static inline size_t quic_decint_size_diff(uint64_t val)
 
 
 /* Returns the maximum value of a QUIC variable-length integer with <sz> as size */
-static inline size_t quic_max_int(size_t sz)
+static inline uint64_t /*size_t*/ quic_max_int(size_t sz)
 {
 	switch (sz) {
 	case 1:
